up until update

+++ robotworld/src/Compass.hpp
@@ -21,11 +21,11 @@ namespace Model
 		/**
 		 *
 		 */
-		static void setStdDev(double aStdDev) {Compass::stddev = aStdDev;}
+		static void setStdDev(unsigned short aStdDev) {Compass::stddev = aStdDev;}
 		/**
 		 *
 		 */
-		static double getStdDev(){ return stddev;}
+		static unsigned short getStdDev(){ return stddev;}
 
 		/**
 		 * Returns a 1-line description of the object
@@ -39,7 +39,7 @@ namespace Model
 		/**
 		 * Standard deviation of the odometer per 10 pixels
 		 */
-		static double stddev;
+		static unsigned short stddev;
 
 	};
